release: fix the 0.3.1 CI and release-gate failures (#1526)
Fixes every red workflow so the 0.3.1 release gate can run and publish v0.3.1.

Release gate: the flaky time_bench_demo clock race on Linux, and on Windows
the tcc non-constant initializer (57 tests), the E070 unreachable ret in
std/term (5), pre-1970 UTC fields via the MSVC CRT (time_parse_demo), and the
Vista condvar API missing from tcc's kernel32.def (net_loopback).

Also: LLVM self-host gate (interfaces owned by another package), stale MCP
stdlib index, all nine Super-Linter linters, prek, and the leak ratchet
re-baselined for the 4.4x bigger stdlib.

Runner/Cloudflare stays red - it needs CLOUDFLARE_API_TOKEN rotated.

+++ benchmark/programs/41_matrix_stdlib_smallops/prog.php
@@ -6,7 +6,11 @@ for ($r = 0; $r < $reps; $r++) {
     $a = array_fill(0, 4, array_fill(0, 4, 0.0));
     for ($i = 0; $i < 4; $i++) {
         for ($j = 0; $j < 4; $j++) {
-            $a[$i][$j] = ($r + $i * 3 + $j * 7 + 1) % 97;
+            // Cast: the matrix is float (array_fill seeds it with 0.0, and the
+            // NumPy twin of this benchmark works in float64), but % yields an
+            // int, which left the array int|float and every product mixing the
+            // two.
+            $a[$i][$j] = (float) (($r + $i * 3 + $j * 7 + 1) % 97);
         }
     }
     $c = array_fill(0, 4, array_fill(0, 4, 0.0));
